requirements.txt
APScheduler==3.10.0
homie-helpers~=0.0.6
python-miio~=0.5.12
service-bootstrap~=0.0.4